split off preprocess
